# libcontainer/cgroups/fs: remove todo since strings.Fields performs well 