@cyphar 2024-09-15
Would you mind writing a small benchmark to see if the extra splits are actually faster?