@cyphar 2024-09-15
Thanks for the patch! Would you mind writing a small benchmark to see if the extra splits aren't slower? (Though I guess this code isn't called more than a couple of times, so maybe it's not that big of a deal...)