@Stavrospanakakis 2024-09-16
@cyphar I created the 2 functions as `func getCpuUsageBreakdownWithSplitN() error` and `func getCpuUsageBreakdownWithFields() error` and mocked all the unnecessary parts.

Using `strings.SplitN` is a bit slower than using `strings.Fields`, but I believe using `strings.SplitN` makes the code more stable and clean.

What do you think?

**Output:**
```
$ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/stavrospanakakis/benchtest
BenchmarkGetCpuUsageBreakdownWithFields-10      15189668                79.04 ns/op
BenchmarkGetCpuUsageBreakdownWithSplitN-10      11211726               105.7 ns/op
PASS
ok      github.com/stavrospanakakis/benchtest   4.155s
```

```go
// bench.go
package bench

import (
	"errors"
	"strconv"
	"strings"
)

// Returns user and kernel usage breakdown in nanoseconds.
func getCpuUsageBreakdownWithSplitN() error {
	const (
		userField   = "user"
		systemField = "system"
	)

	// Expected format:
	// user <usage in ticks>
	// system <usage in ticks>
	data := "user 452278264\nsystem 291429664"

	lines := strings.SplitN(data, "\n", 2)
	if len(lines) < 2 {
		return errors.New("unexpected content in cgroup cpuacct.stat")
	}

	userLine := lines[0]
	userLineFields := strings.SplitN(userLine, " ", 2)
	if len(userLineFields) != 2 || userLineFields[0] != userField {
		return errors.New("unexpected content in cgroup cpuacct.stat")
	}

	userUsageInTicks := strings.TrimSpace(userLineFields[1])
	if _, err := strconv.ParseUint(userUsageInTicks, 10, 64); err != nil {
		return err
	}

	systemLine := lines[1]
	systemLineFields := strings.SplitN(systemLine, " ", 2)
	if len(systemLineFields) != 2 || systemLineFields[0] != systemField {
		return errors.New("unexpected content in cgroup cpuacct.stat")
	}

	systemUsageInTicks := strings.TrimSpace(systemLineFields[1])
	if _, err := strconv.ParseUint(systemUsageInTicks, 10, 64); err != nil {
		return errors.New("unexpected content in cgroup cpuacct.stat")
	}
	return nil
}

// Returns user and kernel usage breakdown in nanoseconds.
func getCpuUsageBreakdownWithFields() error {
	const (
		userField   = "user"
		systemField = "system"
	)

	// Expected format:
	// user <usage in ticks>
	// system <usage in ticks>
	data := "user 452278264\nsystem 291429664"

	// TODO: use strings.SplitN instead.
	fields := strings.Fields(data)
	if len(fields) < 4 || fields[0] != userField || fields[2] != systemField {
		return errors.New("unexpected content in cgroup cpuacct.stat")
	}

	if _, err := strconv.ParseUint(fields[1], 10, 64); err != nil {
		return err
	}

	if _, err := strconv.ParseUint(fields[3], 10, 64); err != nil {
		return err
	}

	return nil
}
```

```go
// bench_test.go
package bench

import "testing"

func BenchmarkGetCpuUsageBreakdownWithFields(b *testing.B) {
	for i := 0; i < b.N; i++ {
		getCpuUsageBreakdownWithFields()
	}
}

func BenchmarkGetCpuUsageBreakdownWithSplitN(b *testing.B) {
	for i := 0; i < b.N; i++ {
		getCpuUsageBreakdownWithSplitN()
	}
}
```