@kolyshkin 2024-09-16
Also, I barely remember that strings.Fields was slower than strings.SplitN back in the day, but now this is fixed so maybe we just need to remove the TODO (after checking that the above is true), or check if using strings.Cut makes the code easier and/or faster, and switch to it.