@Stavrospanakakis 2024-09-16
> @Stavrospanakakis do you mind rewriting the benchmark using https://pkg.go.dev/testing#hdr-Benchmarks and making it part of the PR?

@kolyshkin I added the benchmark test as a part of this pull request. Also, after running the test, `strings.Cut` is faster than `strings.Fields`.

```
go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/stavrospanakakis/benchtest
BenchmarkGetCpuUsageBreakdownWithFields-10    	14914494	        78.71 ns/op
BenchmarkGetCpuUsageBreakdownWithCut-10       	24691188	        48.50 ns/op
PASS
ok  	github.com/stavrospanakakis/benchtest	3.767s
```