@kolyshkin 2024-09-17
I used the benchmark (slightly modified, see below) and the code from this PR, and the gain looks much less dramatic (

```console
[kir@kir-tp1 fs]$ perflock go test -run 1234 -bench . -benchmem -count 8 . > before
[kir@kir-tp1 fs]$ vim cpuacct.go 
[kir@kir-tp1 fs]$ perflock go test -run 1234 -bench . -benchmem -count 8 . > after
[kir@kir-tp1 fs]$ benchstat before after 
goos: linux
goarch: amd64
pkg: github.com/opencontainers/runc/libcontainer/cgroups/fs
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
                        │   before    │               after               │
                        │   sec/op    │   sec/op     vs base              │
GetCpuUsageBreakdown-20   3.751µ ± 3%   3.589µ ± 2%  -4.32% (p=0.002 n=8)

                        │    before    │               after                │
                        │     B/op     │     B/op      vs base              │
GetCpuUsageBreakdown-20   1.945Ki ± 0%   1.883Ki ± 0%  -3.21% (p=0.000 n=8)

                        │   before    │               after               │
                        │  allocs/op  │ allocs/op   vs base               │
GetCpuUsageBreakdown-20   10.000 ± 0%   9.000 ± 0%  -10.00% (p=0.000 n=8)
```

So, given the performance difference and code complexity, we should probably keep strings.Fields as is, remove TODO, add a benchmark and describe in the commit message that strings.Fields gives us a decent performance.