@Stavrospanakakis 2024-09-18
> It would be good to describe in a commit message what we learned here. Something like
> 
> ```
> Initially, this was a commit to switch from strings.Fields to
> strings.SplitN in getCpuUsageBreakdown, but we learned that:
> 
>  - strings.Fields was probably slower than strings.Split[N] in
>    some old Go versions;
>  - it's performance is now adequate (using the new benchmark on versions
>    using strings.SplitN and strings.Cut shows there is <5% performance gain);
> 
> So, remove the TODO (and keep the benchmark).
> ```
> 
> This is just an example to give you an idea of what would make sense to have in a commit message. If you like, you can give it some more time and actually show the benchmark differences between two (or three) versions.

@kolyshkin Sure, I updated the commit message. I also updated the pull request description to include the benchmark differences and the different implementations.